added upload files and update table product list

+++ auto_rti_api/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render
 from rest_framework import viewsets
 from rest_framework.views import APIView
 from rest_framework.response import Response
+from rest_framework.parsers import MultiPartParser
 
 import csv
 from django.db.models import Q
@@ -17,6 +18,8 @@ class WebsitePageViewSet(viewsets.ModelViewSet):
 
 
 class ProductListView(APIView):
+    parser_classes = (MultiPartParser, )
+
     def get(self, request):
         category = request.GET.get('category')
         car_name = request.GET.get('car_name')
@@ -28,18 +31,29 @@ class ProductListView(APIView):
         return Response(product_list_serializer.data)
 
 
-    def post(self, request):
-        ProductList.objects.all().delete()
-        data = csv.reader(open('./auto_rti_api/uploads/price/price.csv'), delimiter=';')
-        for row in data:
-            if row[0] != 'Код':
-                product_list = ProductList()
-                product_list.product_id = row[0]
-                product_list.catalog_id = row[1]
-                product_list.product_name = row[2]
-                product_list.price = row[3]
-                product_list.category = row[7]
-                product_list.car_name = row[8]
-                product_list.manufacturer = row[10]
-                product_list.save()
-        return Response(status = 201)
+    def post(self, request, *args, **kwargs):
+        up_file = request.FILES['file']
+        if up_file.name.split('.')[1] == 'csv':
+            destination = open('./auto_rti_api/uploads/price/' + up_file.name, 'wb+')
+        else:
+            destination = open('./auto_rti_api/uploads/images/' + up_file.name, 'wb+')
+        for chunk in up_file.chunks():
+            destination.write(chunk)
+            destination.close()
+
+        if up_file.name.split('.')[1] == 'csv':
+            ProductList.objects.all().delete()
+            data = csv.reader(open('./auto_rti_api/uploads/price/price.csv'), delimiter=';')
+            for row in data:
+                if row[0] != 'Код':
+                    product_list = ProductList()
+                    product_list.product_id = row[0]
+                    product_list.catalog_id = row[1]
+                    product_list.product_name = row[2]
+                    product_list.price = row[3]
+                    product_list.category = row[7]
+                    product_list.car_name = row[8]
+                    product_list.manufacturer = row[10]
+                    product_list.save()
+        print('# File {} downloaded'.format(up_file.name))
+        return Response({'status': 201})
